refacto: move ac_scan in scanner module
